# app/swagger/api-docs.json
{
	"swagger": "2.0",
	"info": {
		"version": "1.0.0",
		"title": "Swagger CRUD",
		"description": "Swagger CRUD description",
		"termsOfService": "http://swagger.io/terms/"
	},
	"basePath": "/",
	"tags": [
		{
			"name": "Users",
			"description": "User actions"
		}
	],
	"schemes": ["http", "https"],
	"consumes": ["application/json"],
	"produces": ["application/json"],
	"paths": {
		"/api/v1/users": {
			"post": {
				"tags": ["Users"],
				"description": "Create user",
				"produces": ["application/json"],
				"parameters": [
					{
						"name": "vm",
						"required": ["user_id", "username", "password"],
						"in": "body",
						"properties": {
							"user_id": {
								"$ref": "#/definitions/user_id"
							},
							"username": {
								"$ref": "#/definitions/username"
							},
							"password": {
								"$ref": "#/definitions/password"
							}
						}
					}
				],
				"responses": {
					"200": {
						"description": "Success"
					},
					"400": {
						"description": "Business Exception",
						"schema": {
							"$ref": "#/definitions/business_exception"
						}
					}
				}
			},
			"get": {
				"tags": ["Users"],
				"description": "List users",
				"produces": ["application/json"],
				"parameters": [],
				"responses": {
					"200": {
						"description": "Success"
					},
					"400": {
						"description": "Business Exception",
						"schema": {
							"$ref": "#/definitions/business_exception"
						}
					}
				}
			}
		}
	},
	"definitions": {
		"user_id": {
			"type": "number",
			"description": "User ID (1 = ADMIN, 2 = CLIENT)",
			"example": 1
		},
		"username": {
			"type": "string",
			"description": "User logon",
			"example": "Test User"
		},
		"password": {
			"type": "string",
			"description": "User password",
			"example": "1234"
		},
		"error": {
			"type": "object",
			"description": "Error data",
			"properties": {
				"msg": {
					"type": "string",
					"description": "Error message",
					"example": "User not found"
				}
			}
		},
		"business_exception": {
			"type": "array",
			"description": "Error array",
			"items": {
				"$ref": "#/definitions/error"
			}
		}
	}
}